libssc-sensor-compass: rework

+++ src/libssc-cli.c
@@ -38,7 +38,7 @@ compass_close_cb (SSCSensorCompass *self)
 	return G_SOURCE_REMOVE;
 }
 
-static void compass_measurement (SSCSensorCompass *sensor, gdouble heading, gpointer user_data)
+static void compass_measurement (SSCSensorCompass *sensor, gfloat heading, gpointer user_data)
 {
 	g_printf ("Compass sensor measurement: %f °\n", heading);
 }
@@ -59,7 +59,7 @@ magnetometer_close_cb (SSCSensorMagnetometer *self)
 	return G_SOURCE_REMOVE;
 }
 
-static void magnetometer_measurement (SSCSensorMagnetometer *sensor, gdouble magnetic_field_x, gdouble magnetic_field_y, gdouble magnetic_field_z, gpointer user_data)
+static void magnetometer_measurement (SSCSensorMagnetometer *sensor, gfloat magnetic_field_x, gfloat magnetic_field_y, gfloat magnetic_field_z, gpointer user_data)
 {
 	g_printf ("Magnetometer sensor measurement: X=%f Y=%f Z=%f μT\n", magnetic_field_x, magnetic_field_y, magnetic_field_z);
 }
@@ -80,7 +80,7 @@ accelerometer_close_cb (SSCSensorAccelerometer *self)
 	return G_SOURCE_REMOVE;
 }
 
-static void accelerometer_measurement (SSCSensorAccelerometer *sensor, gdouble accel_x, gdouble accel_y, gdouble accel_z, gpointer user_data)
+static void accelerometer_measurement (SSCSensorAccelerometer *sensor, gfloat accel_x, gfloat accel_y, gfloat accel_z, gpointer user_data)
 {
 	g_printf ("Accelerometer sensor measurement: X=%f Y=%f Z=%f m/s²\n", accel_x, accel_y, accel_z);
 }
@@ -101,7 +101,7 @@ light_close_cb (SSCSensorLight *self)
 	return G_SOURCE_REMOVE;
 }
 
-static void light_measurement (SSCSensorLight *sensor, gdouble intensity, gpointer user_data)
+static void light_measurement (SSCSensorLight *sensor, gfloat intensity, gpointer user_data)
 {
 	g_printf ("Light sensor measurement: %f Lux\n", intensity);
 }
